[EDIT] Cleanup classes structure
