+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.MoviesApp"
         tools:targetApi="31">
         <activity
-            android:name=".features.list.ui.MainActivity"
+            android:name=".features.common.ui.activities.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MoviesApp">
