+++ app/src/main/java/com/example/moviesapp/features/common/ui/activities/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.example.moviesapp.features.common.ui.activities
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.example.moviesapp.features.list.ui.components.MoviesListScreen
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MoviesListScreen()
+        }
+    }
+}
+
+
+
