+++ app/src/main/java/com/example/moviesapp/features/list/ui/components/MoviesList.kt
@@ -1,0 +1,91 @@
+package com.example.moviesapp.features.list.ui.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
+import androidx.compose.material3.Card
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import com.example.moviesapp.features.list.ui.models.Movie
+import com.example.moviesapp.features.list.ui.models.getMovieTemp
+
+@Composable
+fun MoviesList(movies: List<Movie>, modifier: Modifier = Modifier) {
+    LazyVerticalGrid(
+        modifier = modifier,
+        columns = GridCells.Fixed(2),
+        contentPadding = PaddingValues(16.dp),
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+        state = rememberLazyGridState()
+    ) {
+        items(movies) { movie ->
+            MovieItem(movie)
+        }
+
+    }
+}
+
+@Composable
+fun MovieItem(movie: Movie, modifier: Modifier = Modifier) {
+    Card(modifier = modifier) {
+        Column {
+            AsyncImage(
+                model = movie.posterUrl,
+                contentDescription = null,
+                modifier = Modifier
+                    .height(200.dp)
+                    .fillMaxWidth(),
+                contentScale = ContentScale.FillWidth
+            )
+            Text(
+                text = movie.title ?: "",
+                modifier = Modifier.padding(horizontal = 8.dp),
+                style = MaterialTheme.typography.titleMedium,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+            Text(
+                text = movie.releaseDate ?: "",
+                modifier = Modifier
+                    .padding(horizontal = 8.dp)
+                    .padding(bottom = 8.dp),
+                style = MaterialTheme.typography.bodySmall
+            )
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun MoviesListPreview() {
+    MoviesList(
+        listOf(
+            getMovieTemp(),
+            getMovieTemp(),
+            getMovieTemp()
+        )
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun MovieItemPreview() {
+    MovieItem(
+        getMovieTemp()
+    )
+}
